rename top folder to src
