+++ main.py
@@ -1,6 +1,6 @@
-from scraper.scraper.spiders.syosetsu_spider import *
-from scraper.text_files_packing import *
-import scraper.text_files_packing as packing
+from src.scraper.spiders.syosetsu_spider import *
+from src.text_files_packing import *
+import src.text_files_packing as packing
 import os
 
 
@@ -46,10 +46,10 @@ if __name__ == '__main__':
     # Feed the novel name and url to spider and crawl the webpages
 
     novels_urls = [
-        #("Ascendance of a Bookworm - Extra Story", "https://ncode.syosetu.com/n4750dy/", chapter_output_range(3)),
+        ("Ascendance of a Bookworm - Extra Story", "https://ncode.syosetu.com/n4750dy/", chapter_output_range(3)),
         #("My Next Life as a Villainess - All Routes Lead to Doom!", "https://ncode.syosetu.com/n5040ce/", chapter_output_range()),
         #("Homeless Tensei ~Isekai de Jiyuu Sugiru Jikyuu Jisoku Seikatsu", "https://ncode.syosetu.com/n8162dq/", chapter_output_range()),
-        ("The Magical Revolution of the Reincarnated Princess and the Genius Young Lady", "https://ncode.syosetu.com/n8558fh/", chapter_output_range(5)),
+        #("The Magical Revolution of the Reincarnated Princess and the Genius Young Lady", "https://ncode.syosetu.com/n8558fh/", chapter_output_range(5)),
         #("Silent Witch", "https://ncode.syosetu.com/n5194gp/", chapter_output_range()),
         #("Food Travel in the Other World with Ridiculous Ability", "https://ncode.syosetu.com/n2710db/", chapter_output_range()),
         #("Frontier Lord Starts Out With a Population of 0", "https://ncode.syosetu.com/n1980en/", chapter_output_range()),
