+++ requirements.txt
@@ -2,8 +2,12 @@ argon2-cffi==21.1.0
 attrs==21.2.0
 Automat==20.2.0
 backcall==0.2.0
+black==23.3.0
 bleach==4.1.0
+certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
 cffi==1.15.0
+charset-normalizer==3.1.0
+click==8.1.3
 colorama==0.4.4
 constantly==15.1.0
 cryptography==36.0.0
@@ -12,6 +16,7 @@ debugpy==1.5.1
 decorator==5.1.0
 defusedxml==0.7.1
 entrypoints==0.3
+filelock==3.11.0
 h2==3.2.0
 hpack==3.0.0
 hyperframe==5.2.0
@@ -39,16 +44,19 @@ lxml==4.6.4
 MarkupSafe==2.0.1
 matplotlib-inline==0.1.3
 mistune==0.8.4
+mypy-extensions==1.0.0
 nbclient==0.5.9
 nbconvert==6.3.0
 nbformat==5.1.3
 nest-asyncio==1.5.4
 notebook==6.4.6
-packaging==21.3
+packaging==23.1
 pandocfilters==1.5.0
 parsel==1.6.0
 parso==0.8.3
+pathspec==0.11.1
 pickleshare==0.7.5
+platformdirs==3.2.0
 priority==1.3.0
 prometheus-client==0.12.0
 prompt-toolkit==3.0.23
@@ -68,19 +76,26 @@ pyzmq==22.3.0
 qtconsole==5.2.1
 QtPy==1.11.3
 queuelib==1.6.2
-Scrapy==2.5.1
+requests==2.28.2
+requests-file==1.5.1
+Scrapy==2.8.0
+scrapy-utils==0.0.6
 Send2Trash==1.8.0
 service-identity==21.1.0
 six==1.16.0
 terminado==0.12.1
 testpath==0.5.0
+tldextract==3.4.0
+tomli==2.0.1
 tornado==6.1
 traitlets==5.1.1
 Twisted==21.7.0
 twisted-iocpsupport==1.0.2
 typing_extensions==4.0.1
+urllib3==1.26.15
 w3lib==1.22.0
 wcwidth==0.2.5
 webencodings==0.5.1
 widgetsnbextension==3.5.2
+wincertstore==0.2
 zope.interface==5.4.0
